Convert tests to pytest
